+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+* Add `_clear` for deleting all `enumeratedValues` from field
 * Support for collecting fields in field arrays
 * Deriving fields